+++ R/visualize_helper.R
@@ -33,7 +33,7 @@ Datetime_breaks <- function(x,
 #' Take a vector of `Datetimes` and return the start of the first and end of the
 #' last day of data. The `start` and the `length` can be adjusted by
 #' `durations`, like [lubridate::ddays()]. It is used in the [gg_days()]
-#' function to return a sensible axis. This function is a thin wrapper around
+#' function to return a sensible x-axis. This function is a thin wrapper around
 #' [lubridate::floor_date()] and [lubridate::ceiling_date()].
 #'
 #' @param x a vector of `Datetimes`
@@ -41,9 +41,11 @@ Datetime_breaks <- function(x,
 #'   shifts the start of the `Datetime` vector by this amount.
 #' @param length optional `duration` object, e.g. `lubridate::ddays(7)` that
 #'   shifts the end of the `Datetime` vector by this amount from the (adjusted)
-#'   start.
+#'   start. Depending on the data, you might have to subtract one day from the
+#'   desired length to get the correct axis-scaling if you start at midnight.
 #' @param unit a `character` scalar giving the unit of rounding in
 #'   [lubridate::floor_date()] and [lubridate::ceiling_date()]
+#' @param doubleplot a `logical` scalar indicating if used in [gg_doubleplot()]
 #' @param ... other arguments passed to [lubridate::floor_date()] and
 #'   [lubridate::ceiling_date()]
 #'
@@ -61,6 +63,7 @@ Datetime_limits <- function(x,
                             start = NULL,
                             length = NULL,
                             unit = "1 day",
+                            doubleplot = FALSE,
                             ...) {
   
   min_date <- x %>% lubridate::as_datetime() %>% min()
@@ -75,6 +78,12 @@ Datetime_limits <- function(x,
       y <- (min_date + length)
       y %>% lubridate::as_datetime()
     }
+  
+  if(doubleplot) {
+  if(identical(max_date, lubridate::ceiling_date(max_date, unit = unit, ...))) {
+    max_date <- max_date + lubridate::duration(unit)
+  }
+  }
   
   c(lubridate::floor_date(min_date, unit = unit, ...),
     lubridate::ceiling_date(max_date, unit = unit, ...))